+++ Quick/Control/DataBox.cpp
@@ -265,7 +265,7 @@ void CDataBox::ToExcel(CString strTitle)
 	long cols = GetColumns()->GetVisibleColumnsCount();
 
 	COleSafeArray sa;
-	unsigned long elements[] = {rows, cols};
+	unsigned long elements[] = { static_cast<unsigned long>(rows), static_cast<unsigned long>(cols) };
 	sa.Create(VT_VARIANT, 2, elements);
 
 	for(int r = 0; r < rows; r++)
